rearchitect the app to implement routers and multiple pages
