+++ client/src/pages/Group.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class Group extends Component {
+  render() {
+    return (<div>Group</div>);
+  }
+}
+
+export default Group;
