+++ client/src/pages/Home.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import { browserHistory } from 'react-router';
+
+class Home extends Component {
+  render () {
+    return (
+      <div>
+        Main
+        <button onClick={() => browserHistory.push('/profile/12')}>Click</button>
+      </div>
+    );
+  }
+}
+
+export default Home;
