+++ client/src/pages/NotFound.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class NotFound extends Component {
+  render () {
+    return (<div>NotFound</div>);
+  }
+}
+
+export default NotFound;
