+++ client/src/pages/Profile.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class Profile extends Component {
+  render () {
+    return (<div>Profile{this.props.params.username}</div>);
+  }
+}
+
+export default Profile;
