+++ package.json
@@ -3,19 +3,29 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "npm run build | npm run server",
-    "build": "./node_modules/.bin/gulp build",
-    "server": "nodemon --ignore 'client/*.js' server/server.js",
-    "watch": "./node_modules/.bin/gulp watch",
+    "start": "npm run server",
+    "server": "nodemon ./server/server.js",
     "lint": "./node_modules/.bin/eslint **/*.js **/*.jsx *.js *.jsx",
-    "graphql": "nodemon ./server/tempserver.js"
+    "graphql": "nodemon ./server/tempserver.js",
+    "build": "webpack --watch"
   },
   "devDependencies": {
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.0.0",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "css-loader": "^0.14.5",
     "eslint": "^3.19.0",
     "eslint-config-standard": "^10.2.1",
+    "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-node": "^4.2.2",
     "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-standard": "^3.0.1"
+    "eslint-plugin-standard": "^3.0.1",
+    "extract-text-webpack-plugin": "^2.0.0-beta.4",
+    "style-loader": "^0.16.1",
+    "webpack": "^2.4.1"
   },
   "dependencies": {
     "axios": "^0.15.3",
@@ -23,7 +33,6 @@
     "babel-plugin-transform-react-jsx": "^6.8.0",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
-    "babelify": "^7.3.0",
     "body-parser": "^1.15.2",
     "bootstrap": "^3.3.7",
     "browserify": "^13.1.1",
@@ -34,9 +43,6 @@
     "express-graphql": "^0.6.4",
     "express-session": "^1.14.2",
     "graphql": "^0.9.3",
-    "gulp": "^3.9.1",
-    "gulp-sourcemaps": "^1.9.1",
-    "jquery": "^3.1.1",
     "lodash": "^4.17.2",
     "mocha": "^3.2.0",
     "mongoose": "^4.7.1",
@@ -47,12 +53,10 @@
     "passport-facebook": "^2.1.1",
     "react": "^15.4.1",
     "react-dom": "^15.4.1",
-    "react-faux-dom": "^3.0.0",
     "react-modal-bootstrap": "^1.1.0",
+    "react-router": "^3.0.2",
+    "react-router-dom": "^4.1.1",
     "socket.io": "^1.7.1",
-    "socket.io-client": "^1.7.1",
-    "vinyl-buffer": "^1.0.0",
-    "vinyl-source-stream": "^1.1.0",
-    "watchify": "^3.7.0"
+    "socket.io-client": "^1.7.1"
   }
 }
