+++ server/models/user.js
@@ -1,0 +1,7 @@
+const db = require('../db/schemas');
+
+module.exports = {
+  getInfo: (userId) => {
+    return db.User.findOne({ _id: userId });
+  }
+};
